requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
snscrape>=0.7.0
apscheduler>=3.10.0
python-dateutil>=2.8.0
nltk>=3.8.0
matplotlib>=3.7.0
